requirements.txt
streamlit
tensorflow
numpy
pillow
seaborn
scikit-learn
matplotlib
pandas
